Split VipsConverter into separate image/vips/converter module
Move VipsConverter from image/vips/deployment to image/vips/converter
so it can be used at runtime by extensions like Roq for on-the-fly
image processing without depending on the deployment module.
